+++ lib/datastore/README.md
@@ -1,0 +1,5 @@
+# mapeo-core datastore
+
+## Purpose
+
+## Usage
