+++ lib/datatype/README.md
@@ -1,0 +1,5 @@
+# mapeo-core datatype
+
+## Purpose
+
+## Usage
