+++ lib/indexer/README.md
@@ -1,0 +1,5 @@
+# mapeo-core indexer
+
+## Purpose
+
+## Usage
