+++ lib/types.js
@@ -45,6 +45,14 @@
  */
 
 /**
+ * @typedef {Object<string, any>} Doc
+ */
+
+/**
+ * @typedef {Buffer} Block
+ */
+
+/**
  * @typedef {Object} NoiseSecretStream
  * @property {Buffer} remotePublicKey
  * @property {number} remotePort
