+++ package.json
@@ -31,16 +31,17 @@
   "homepage": "https://github.com/digidem/mapeo-core#readme",
   "devDependencies": {
     "@hyperswarm/testnet": "^3.1.0",
-    "@types/node": "^18.7.6",
+    "@types/json-schema": "^7.0.11",
+    "@types/node": "^18.7.15",
     "@types/streamx": "^2.9.1",
-    "brittle": "^3.0.4",
+    "brittle": "^3.1.0",
     "dependency-check": "^4.1.0",
-    "eslint": "^8.22.0",
+    "eslint": "^8.23.0",
     "eslint-config-prettier": "^8.5.0",
     "hypercore-crypto": "^3.3.0",
     "prettier": "^2.7.1",
-    "random-access-memory": "^5.0.0",
-    "typescript": "^4.7.4"
+    "random-access-memory": "^5.0.1",
+    "typescript": "^4.8.2"
   },
   "dependencies": {
     "@hyperswarm/secret-stream": "^6.0.0",
@@ -49,12 +50,11 @@
     "b4a": "^1.6.0",
     "base32.js": "^0.1.0",
     "better-sqlite3": "^7.6.2",
-    "corestore": "^6.0.0",
-    "hypercore": "^10.0.0",
+    "corestore": "^6.0.4",
+    "hypercore": "^10.2.1",
     "hyperswarm": "^4.2.0",
     "multi-core-indexer": "github:digidem/multi-core-indexer",
     "multicast-service-discovery": "^4.0.3",
-    "mutexify": "^1.4.0",
     "sodium-native": "^3.4.1",
     "tiny-typed-emitter": "^2.1.0"
   }
